move remote tests into a new file
I was not able to register a mock resource for these tests for some
reason. They will continue using implemented resources for now.
